Get tibbles from Hydroscope

A family of functions that return a tibble from a specific database from Hydroscope using the Enhydris API. get_database returns a named list of tibbles using all the family's functions.

get_stations(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_timeseries(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_instruments(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_water_basins(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_water_divisions(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_political_divisions(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_variables(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_units_of_measurement(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_time_steps(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_owners(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

get_instruments_type(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_station_type(subdomain = c("kyy", "ypaat", "emy", "deh"),
  translit = TRUE)

get_database(subdomain = c("kyy", "ypaat", "emy", "deh"), translit = TRUE)

Arguments

subdomain
One of the subdomains of Hydroscope in the vector c("kyy", "ypaat", "emy", "deh").

translit
Automatically transliterate Greek to Latin.

Value

If subdomain is one of:

kyy, Ministry of Environment and Energy.

ypaat, Ministry of Rural Development and Food.

deh, Greek Public Power Corporation.

emy, National Meteorological Service.

returns a tibble or a named list with tibbles from the corresponding database. Otherwise returns an error message.

Note

Objects' IDs are not unique among the different Hydroscope databases. For example, time series' IDs from http://kyy.hydroscope.gr have same values with time series' from http://ypaat.hydroscope.gr.

The coordinates of the stations are based on the European Terrestrial Reference System 1989 (ETRS89).

References

The data are retrieved from the Hydroscope's databases:

Ministry of Environment, Energy and Climate Change, http://kyy.hydroscope.gr

Ministry of Rural Development and Food, http://ypaat.hydroscope.gr

National Meteorological Service, http://emy.hydroscope.gr

Greek Public Power Corporation, http://deh.hydroscope.gr

European Terrestrial Reference System 1989 (ETRS), http://bit.ly/2kJwFuf

Greek Water Divisions, http://bit.ly/2kk0tOm, http://bit.ly/2ltQC8O

Greek Water Basins, http://bit.ly/2Dvzo1W

Tibble, http://tibble.tidyverse.org/

Examples

# NOT RUN {
# data will be downloaded from Ministry of Environment and Energy (kyy):
subdomain <- "kyy"

# stations
kyy_stations <- get_stations(subdomain)

# time series
kyy_ts <- get_timeseries(subdomain)

# instruments
kyy_inst <- get_instruments(subdomain)

# water basins
kyy_wbas <- get_water_basins(subdomain)

# water divisions
kyy_wdiv <- get_water_divisions(subdomain)

# political divisions
kyy_pol <- get_political_divisions(subdomain)

# variables
kyy_vars <- get_variables(subdomain)

# units of measurement
kyy_units <- get_units_of_measurement(subdomain)

# time steps
kyy_time_steps <- get_time_steps(subdomain)

# owners
kyy_owners <- get_owners(subdomain)

# instruments type
kyy_instr_type <- get_instruments_type(subdomain)

# stations' type
kyy_st_type <- get_station_type(subdomain)

# use all the get_ functions above to create a named list with tibbles
kyy_db <- get_database(subdomain)

# }