Translate Greek names and terms to English

hydro_translate Greeek to English translations of names and terms

hydro_translate(x, value = c("owner", "variable", "timestep", "division"))

Arguments

x
a string vector

value
One of the predefined values in c("owner", "variable", "timestep", "division")

Value

If value is one of:

owner, translate organisations' names

variable, translate hydrometeorological term

returns a character vector with translations of various hydrometeorological terms or organizations' names

Examples

# NOT RUN {
# get data from the Ministry of Environment and Energy
kyy_owners <- get_owners("kyy")
kyy_vars <- get_variables("kyy")
owners_names <- hydro_translate(kyy_owners$name, "owner")
vars <- hydro_translate(kyy_vars$descr, "variable")
# }